Made suggested fix
